move normal functor to random_functor.cpp
