+++ oneflow/core/functional/impl/nn_functor.cpp
@@ -26,8 +26,6 @@ limitations under the License.
 #include "oneflow/user/kernels/dropout_kernel.h"
 #include "oneflow/core/common/container_util.h"
 #include "oneflow/user/kernels/distributions/common.h"
-#include "oneflow/user/kernels/random_seed_util.h"
-#include "oneflow/core/rpc/include/global_process_ctx.h"
 
 namespace oneflow {
 namespace one {
@@ -2001,168 +1999,6 @@ class GridSampleFunctor {
 
  private:
   std::shared_ptr<OpExpr> op_;
-};
-
-class NormalFunctor {
- public:
-  NormalFunctor() { op_ = CHECK_JUST(one::OpBuilder("normal").Output("out").Build()); }
-  Maybe<Tensor> operator()(const float& mean, const float& std, const Shape& shape,
-                           const Optional<one::Tensor>& out,
-                           const Optional<Symbol<DType>>& optional_dtype,
-                           const Optional<Symbol<Device>>& optional_device,
-                           const Optional<one::Generator>& optional_generator,
-                           const bool& requires_grad) const {
-    Symbol<DType> dtype = DType::Float();
-    if (optional_dtype.has_value()) {
-      dtype = JUST(optional_dtype);
-      if (dtype->data_type() != DataType::kFloat && dtype->data_type() != DataType::kDouble) {
-        OF_UNIMPLEMENTED() << "Only support float and double in normal().";
-      }
-    }
-    Symbol<Device> device = JUST(Device::New("cpu"));
-    if (optional_device.has_value()) { device = JUST(optional_device); }
-
-    if (out.has_value()) {
-      auto out_tensor = JUST(out);
-      Symbol<DType> output_tensor_dtype = out_tensor->dtype();
-      if (optional_dtype.has_value()) {
-        CHECK_OR_RETURN(output_tensor_dtype == dtype)
-            << Error::RuntimeError() << "data type " << dtype->name()
-            << " does not match data type of out parameter " << output_tensor_dtype->name();
-      }
-      dtype = output_tensor_dtype;
-      Symbol<Device> out_tensor_device = JUST(out_tensor->device());
-      if (optional_device.has_value()) {
-        CHECK_OR_RETURN(out_tensor_device == JUST(optional_device))
-            << Error::RuntimeError() << "device type " << device->ToString()
-            << " does not match device type of out parameter " << out_tensor_device->ToString();
-      }
-      device = out_tensor_device;
-    }
-    const auto gen = optional_generator.value_or(JUST(one::DefaultAutoGenerator()));
-    auto& attrs = THREAD_CACHED_MUTABLE_ATTR_MAP("mean", "std", "shape", "dtype", "seed");
-    attrs.SetAllAttrs(static_cast<double>(mean), static_cast<double>(std), shape,
-                      dtype->data_type(), static_cast<int64_t>(gen->current_seed()));
-
-    const auto& distribution_state = std::make_shared<DistributionKernelState>(gen);
-    OpExprInterpContext ctx(attrs, distribution_state);
-    ctx.device = device;
-    if (out.has_value()) {
-      std::shared_ptr<TensorTuple> outputs = std::make_shared<TensorTuple>(1);
-      (*outputs)[0] = JUST(out);
-      JUST(OpInterpUtil::Dispatch(*op_, {}, outputs.get(), ctx));
-      return (*outputs)[0];
-    }
-
-    auto result = JUST(OpInterpUtil::Dispatch<Tensor>(*op_, {}, ctx));
-    JUST(result->set_requires_grad(requires_grad));
-    return result;
-  }
-
- private:
-  std::shared_ptr<OpExpr> op_;
-};
-
-class Normal2Functor {
- public:
-  Maybe<Tensor> operator()(const float& mean, const float& std, const int32_t& shape,
-                           const Optional<one::Tensor>& out,
-                           const Optional<Symbol<DType>>& optional_dtype,
-                           const Optional<Symbol<Device>>& optional_device,
-                           const Optional<one::Generator>& optional_generator,
-                           const bool& requires_grad) const {
-    const Shape size = Shape({shape});
-    return Normal(mean, std, size, out, optional_dtype, optional_device, optional_generator,
-                  requires_grad);
-  }
-};
-
-class GlobalNormalFunctor {
- public:
-  GlobalNormalFunctor() { op_ = CHECK_JUST(one::OpBuilder("normal").Output("out").Build()); }
-  Maybe<Tensor> operator()(const float& mean, const float& std, const Shape& shape,
-                           const Optional<one::Tensor>& out, const Symbol<ParallelDesc>& placement,
-                           const std::vector<Symbol<SbpParallel>>& sbp_tuple,
-                           const Optional<Symbol<DType>>& optional_dtype,
-                           const Optional<one::Generator>& optional_generator,
-                           const bool& requires_grad) const {
-    JUST(CheckDeviceIdsIsValid(placement));
-
-    Symbol<DType> dtype = DType::Float();
-    if (optional_dtype.has_value()) {
-      dtype = JUST(optional_dtype);
-      if (dtype->data_type() != DataType::kFloat && dtype->data_type() != DataType::kDouble) {
-        OF_UNIMPLEMENTED() << "Only support float and double in normal().";
-      }
-    }
-
-    if (out.has_value()) {
-      auto out_tensor = JUST(out);
-      Symbol<DType> output_tensor_dtype = out_tensor->dtype();
-      if (optional_dtype.has_value()) {
-        CHECK_OR_RETURN(output_tensor_dtype == dtype)
-            << Error::RuntimeError() << "data type " << dtype->name()
-            << " does not match data type of out parameter (" << output_tensor_dtype->name();
-      }
-      dtype = output_tensor_dtype;
-    }
-
-    auto& attrs = THREAD_CACHED_MUTABLE_ATTR_MAP("mean", "std", "shape", "dtype", "seed", "nd_sbp");
-
-    const auto& nd_sbp = JUST(GetNdSbp(sbp_tuple));
-
-    std::shared_ptr<Generator> gen = optional_generator.value_or(JUST(one::DefaultAutoGenerator()));
-    uint64_t init_seed = JUST(gen->Get<CPUGeneratorImpl>(0))->engine()();
-
-    if (LazyMode::is_enabled()) {
-      attrs.SetAllAttrs(static_cast<double>(mean), static_cast<double>(std), shape,
-                        dtype->data_type(), static_cast<int64_t>(init_seed),
-                        *JUST(GetNdSbpStrList(nd_sbp)));
-    } else {
-      uint64_t rank_seed = 0;
-      {
-        JUST(BroadcastSeedToAllRanks(&init_seed, /*root=*/0));
-        rank_seed =
-            JUST(GetRandomSeedForRank(*placement, *nd_sbp, init_seed, GlobalProcessCtx::Rank()));
-      }
-      attrs.SetAllAttrs(static_cast<double>(mean), static_cast<double>(std), shape,
-                        dtype->data_type(), static_cast<int64_t>(rank_seed), NullOpt);
-      gen = JUST(MakeGenerator(placement->device_type()));
-      gen->set_current_seed(rank_seed);
-    }
-    const auto& distribution_state = std::make_shared<DistributionKernelState>(gen);
-
-    if (out.has_value()) {
-      std::shared_ptr<TensorTuple> outputs = std::make_shared<TensorTuple>(1);
-      (*outputs)[0] = JUST(out);
-      JUST(OpInterpUtil::Dispatch(
-          *op_, {}, outputs.get(),
-          OpExprInterpContext(attrs, placement, nd_sbp, distribution_state)));
-      return (*outputs)[0];
-    }
-
-    auto result = JUST(OpInterpUtil::Dispatch<Tensor>(
-        *op_, {}, OpExprInterpContext(attrs, placement, nd_sbp, distribution_state)));
-    JUST(result->set_requires_grad(requires_grad));
-    return result;
-  }
-
- private:
-  std::shared_ptr<OpExpr> op_;
-};
-
-class GlobalNormal2Functor {
- public:
-  Maybe<Tensor> operator()(const float& mean, const float& std, const int32_t& shape,
-                           const Optional<one::Tensor>& out, const Symbol<ParallelDesc>& placement,
-                           const std::vector<Symbol<SbpParallel>>& sbp_tuple,
-                           const Optional<Symbol<DType>>& optional_dtype,
-                           const Optional<one::Generator>& optional_generator,
-                           const bool& requires_grad) const {
-    const Shape size = Shape({shape});
-    return GlobalNormal(mean, std, size, out, placement, sbp_tuple, optional_dtype,
-                        optional_generator, requires_grad);
-  }
 };
 
 class NormalizationFunctor {
@@ -4576,10 +4412,6 @@ ONEFLOW_FUNCTION_LIBRARY(m) {
   m.add_functor<impl::OneEmbeddingFusedLookupGradFunctor>("OneEmbeddingFusedLookupGrad");
   m.add_functor<impl::OneEmbeddingEmbeddingPutFunctor>("OneEmbeddingEmbeddingPut");
   m.add_functor<impl::OneEmbeddingUniqueKeyValuePairFunctor>("OneEmbeddingUniqueKeyValuePair");
-  m.add_functor<impl::NormalFunctor>("Normal");
-  m.add_functor<impl::Normal2Functor>("Normal2");
-  m.add_functor<impl::GlobalNormalFunctor>("GlobalNormal");
-  m.add_functor<impl::GlobalNormal2Functor>("GlobalNormal2");
   m.add_functor<impl::OneEmbeddingSgdUpdateFunctor>("OneEmbeddingSgdUpdate");
   m.add_functor<impl::OneEmbeddingAdamUpdateFunctor>("OneEmbeddingAdamUpdate");
   m.add_functor<impl::OneEmbeddingAdagradUpdateFunctor>("OneEmbeddingAdagradUpdate");
